Files separation for same package
