+++ master/main.go
@@ -6,7 +6,6 @@ import (
 	"log"
 	"net/http"
 	"strings"
-	"time"
 
 	"github.com/gorilla/websocket"
 	"gopkg.in/mgo.v2"
@@ -24,39 +23,6 @@ var (
 		WriteBufferSize: 1024,
 	}
 )
-
-// User model prototype without hashing algorithms implemented yet
-type User struct {
-	ID           bson.ObjectId `bson:"_id,omitempty"`
-	Username     string        `bson:"username,omitempty"`
-	Password     string        `bson:"password,omitempty"`
-	FirstName    string        `bson:"firstName"`
-	SecondName   string        `bson:"secondName"`
-	RegisterDate time.Time     `bson:"registerDate"`
-}
-
-//StorageServer represents a Network Attached Storage device
-type StorageServer struct {
-	ID   bson.ObjectId `bson:"_id,omitempty"`
-	Name string        `bson:"name"`
-}
-
-//BlockDevice represents a block device retrieved by blkid probe
-type BlockDevice struct {
-	ID    bson.ObjectId `bson:"_id,omitempty"`
-	VolID bson.ObjectId `bson:"volID"` //can be empty
-	Path  string        `bson:"path,omitempty"`
-	UUID  string        `bson:"uuid,omitempty"`
-	Type  string        `bson:"type,omitempty"`
-}
-
-//BtrfsVolume represents a filesystem volume which can potentially span over
-//multiple devices
-type BtrfsVolume struct {
-	ID     bson.ObjectId `bson:"_id,omitempty"`
-	ServID bson.ObjectId `bson:"servID"` // can be empty
-	Label  string        `bson:"label"`
-}
 
 func authentication(loginAndPass []string) bool {
 	result := User{}
